agent/pyproject.toml
[project]
name = "ai-charactermancer"
version = "0.1.0"
description = "AI sidekick agent for Pathfinder 1e character creation."
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "google-genai>=1.0",
    "langchain-mongodb>=0.11",
    "langchain-google-genai>=2.0",
    "pymongo>=4.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_charactermancer"]

[tool.ruff.format]
preview = true

[tool.ruff.lint]
preview = true